Move Taffy-specific stuff into taffy_tree module
