+++ Cargo.toml
@@ -22,7 +22,7 @@ slotmap = "1.0.6"
 grid = { version = "0.9.0", optional = true }
 
 [features]
-default = ["std", "flexbox", "grid"]
+default = ["std", "flexbox", "grid", "taffy"]
 flexbox = []
 grid = ["alloc", "dep:grid"]
 alloc = []
