+++ src/tree/taffy_tree/mod.rs
@@ -1,0 +1,6 @@
+mod leaf;
+mod tree;
+mod error;
+
+pub use tree::{Taffy, TaffyChildIter};
+pub use error::{TaffyError, TaffyResult};
